refactor(Score): ChessBoard 점수 계산 로직 Score 클래스 분리
