+++ src/main/java/domain/chessgame/ChessBoard.java
@@ -1,20 +1,14 @@
 package domain.chessgame;
 
-import domain.Player;
 import domain.piece.Blank;
 import domain.piece.Piece;
-import domain.position.File;
 import domain.position.Position;
-import domain.position.Rank;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 public class ChessBoard {
 
     private static final int DEFAULT_KING_COUNT = 2;
-    private static final int SEVERAL_PAWN_COUNT = 2;
     private static final int PAWN_ATTACK = 0;
 
     private final Map<Position, Piece> board;
@@ -62,34 +56,6 @@ public class ChessBoard {
         Piece sourcePiece = board.get(source);
         board.put(target, sourcePiece);
         board.put(source, new Blank());
-    }
-
-    public double calculateScoreByPlayer(final Player player) {
-        return Arrays.stream(File.values())
-            .map(file -> calculateTotalScoreInFile(player, file))
-            .mapToDouble(Double::doubleValue)
-            .sum();
-    }
-
-    private double calculateTotalScoreInFile(final Player player, final File file) {
-        List<Piece> pieces = createPiecesInFile(player, file);
-        boolean isSeveralPawn = isSeveralPawnInList(pieces);
-        return pieces.stream().map(piece -> piece.score(isSeveralPawn))
-            .mapToDouble(Double::doubleValue)
-            .sum();
-    }
-
-    private List<Piece> createPiecesInFile(final Player player, final File file) {
-        return Arrays.stream(Rank.values())
-            .map(rank -> board.get(Position.of(file, rank)))
-            .filter(piece -> piece.isSamePlayer(player))
-            .collect(Collectors.toList());
-    }
-
-    private boolean isSeveralPawnInList(final List<Piece> pieces) {
-        return pieces.stream()
-            .filter(Piece::isPawn)
-            .count() >= SEVERAL_PAWN_COUNT;
     }
 
     public boolean isKingOnlyOne() {
